support transitions > 65s

+++ wled00/wled.h
@@ -591,8 +591,8 @@ WLED_GLOBAL uint8_t paletteBlend _INIT(0);        // determines blending and wra
 // transitions
 WLED_GLOBAL uint8_t       blendingStyle            _INIT(0);      // effect blending/transitionig style
 WLED_GLOBAL bool          transitionActive         _INIT(false);
-WLED_GLOBAL uint16_t      transitionDelay          _INIT(750);    // global transition duration
-WLED_GLOBAL uint16_t      transitionDelayDefault   _INIT(750);    // default transition time (stored in cfg.json)
+WLED_GLOBAL uint32_t      transitionDelay          _INIT(750);    // global transition duration (in ms, max. TRANSITION_MAX_DUR)
+WLED_GLOBAL uint32_t      transitionDelayDefault   _INIT(750);    // default transition time (in ms, stored in cfg.json)
 WLED_GLOBAL unsigned long transitionStartTime;
 WLED_GLOBAL bool          jsonTransitionOnce       _INIT(false);  // flag to override transitionDelay (playlist, JSON API: "live" & "seg":{"i"} & "tt")
 WLED_GLOBAL uint8_t       randomPaletteChangeTime  _INIT(5);      // amount of time [s] between random palette changes (min: 1s, max: 255s)
